# bureau/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for the Bureau MCP Server.",
        "title": "Bureau API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/address": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Get a list of all addresses.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/address"
                ],
                "summary": "List all addresses",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit the number of results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.PaginatedResponse-github_com_BrunoPolaski_bureau_internal_infra_controllers_dto_AddressDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Creates a new address with the provided details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/address"
                ],
                "summary": "Create a new address",
                "parameters": [
                    {
                        "description": "Address",
                        "name": "address",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.AddressDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.AddressDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Delete an address by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/address"
                ],
                "summary": "Delete an address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Address ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            }
        },
        "/address/{id}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Get an address by its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/address"
                ],
                "summary": "Get an address by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Address ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.AddressDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            }
        },
...
        "/api-key": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Get a list of all apiKeys.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/api-key"
                ],
                "summary": "List all apiKeys",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.PaginatedResponse-github_com_BrunoPolaski_bureau_internal_infra_controllers_dto_ApiKeyDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Creates a new apiKey with the provided details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/api-key"
                ],
                "summary": "Create a new apiKey",
                "parameters": [
                    {
                        "description": "ApiKey",
                        "name": "apiKey",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.ApiKeyDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.ApiKeyDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            }
        },
        "/api-key/{uuid}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Get an apiKey by its UUID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/api-key"
                ],
                "summary": "Get an apiKey by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ApiKey UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.ApiKeyDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Delete a apiKey by its UUID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/api-key"
                ],
                "summary": "Delete a apiKey",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ApiKey UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.AddressDTO": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "complement": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "number": {
                    "type": "string",
                    "example": "1234 | S/N"
                },
                "state": {
                    "type": "string",
                    "example": "SC"
                },
                "street": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "zip_code": {
                    "type": "string",
                    "example": "12345-678"
                }
            }
        },
...
        "github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.ApiKeyDTO": {
            "type": "object",
            "required": [
                "slug",
                "uuid"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                }
            }
        },
...
        "github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.PaginatedResponse-github_com_BrunoPolaski_bureau_internal_infra_controllers_dto_AddressDTO": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "List of items in the current page",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.AddressDTO"
                    }
                },
                "total": {
                    "description": "Total number of items available",
                    "type": "integer"
                }
            }
        },
...
        "github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.PaginatedResponse-github_com_BrunoPolaski_bureau_internal_infra_controllers_dto_ApiKeyDTO": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "List of items in the current page",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_BrunoPolaski_bureau_internal_infra_controllers_dto.ApiKeyDTO"
                    }
                },
                "total": {
                    "description": "Total number of items available",
                    "type": "integer"
                }
            }
        },
...
        "rest_err.Causes": {
            "type": "object",
            "properties": {
                "field": {
                    "description": "Field or parameter that caused the error",
                    "type": "string",
                    "example": "email"
                },
                "message": {
                    "description": "Description of the cause",
                    "type": "string",
                    "example": "invalid email address"
                }
            }
        },
        "rest_err.RestErr": {
            "type": "object",
            "properties": {
                "causes": {
                    "description": "Detailed error causes, most common for json field validation errors",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest_err.Causes"
                    }
                },
                "code": {
                    "description": "HTTP status code",
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "bad request"
                },
                "message": {
                    "description": "Human readable message",
                    "type": "string",
                    "example": "invalid request parameters"
                },
                "timestamp": {
                    "description": "When the error occurred",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-api-key",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "CookieAuth": {
            "type": "apiKey",
            "name": "Sid",
            "in": "cookie"
        }
    }
}